Add concurrency and computeIfAbsent
Modifies hashmaps into ConcurrentMaps / ConcurrentHashMaps so proper concurrency occurs when executing many new instances at once.

ComputeIfAbsent is applied instead of map.put() because of concurrency issues when Camunda was starting up.  When Groovy was executing the first instances of the classes for Simple Metric Classes, the concurrency was being held properly because of the eval if the key was currently in the map.  ComputeIfAbsent ensures the proper atomic lookup is ocurring.